Adds port with resource_parameters
This required the introduction of a pre and post initialize operation.
I was going to leave it at that but then it feels strange to try
and perform work on the args and expect them to return them at the
end of the pre_initialize method. So process_arguments was born.