Adds support to find matchers defined in the FilterTable
This work brought up interesting questions. Like matchers really
are properties used in a particular situation and so it may be
better to define all matchers as properties and add a flag to
to them to denote their "matcher-ness". At the moment the only
matchers that I have come in contact with end with question marks.
I know there are other matchers that start with 'has_...'. So
perhaps they are simply all properties and would be returned as
properties. But there is a filtering of the properties for that
setting or simply make the assumption on prefixes and postfixes.